Move layer state to subfolder
