+++ containers/web-server/frontend/src/details/solutions/SolutionsSidebarContent.tsx
@@ -8,8 +8,8 @@ import { VECTOR_COLOR_MAPS } from '@/config/color-maps';
 import { MARINE_HABITATS_LOOKUP } from '@/config/solutions/domains';
 import { DataItem } from '@/details/features/detail-components';
 import { ColorBox } from '@/map/tooltip/content/ColorBox';
-import { habitatColorMap } from '@/state/layers/marine';
-import { landuseColorMap } from '@/state/layers/terrestrial';
+import { habitatColorMap } from '@/state/layers/data-layers/marine';
+import { landuseColorMap } from '@/state/layers/data-layers/terrestrial';
 
 const slopeColorFunction = colorMap(VECTOR_COLOR_MAPS.terrestrialSlope);
 const elevationColorFunction = colorMap(VECTOR_COLOR_MAPS.terrestrialElevation);
