+++ containers/web-server/frontend/src/state/layers/data-layers/drought.ts
