+++ containers/web-server/frontend/src/state/layers/data-layers/population.ts
@@ -3,8 +3,7 @@ import { selector } from 'recoil';
 import { ViewLayer } from '@/lib/data-map/view-layers';
 
 import { jrcPopulationViewLayer } from '@/config/population/population-view-layer';
-
-import { sectionVisibilityState } from '../sections';
+import { sectionVisibilityState } from '@/state/sections';
 
 export const populationLayerState = selector<ViewLayer>({
   key: 'populationLayerState',
